VitAI/pyproject.toml
[project]
name = "VitAI"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.12.4",
    "azure-ai-inference",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
